projects:ad7768_evb: Use new iio_app files
Old custom app_iio files where used by the project.
Update it to use new ones from noos/iio/iio_app
